feat: Implement user profile editing and OTP verification components
- Added EditProfile component for users to edit their profile information.
- Removed obsolete Login and Register components.
- Introduced CreateAccount component for OTP verification during registration.
- Updated Profile component to display user information without editing functionality.
- Created a new Header component for navigation with view options.
- Removed deprecated Header1 component.
- Defined UserType interface in the instance module for consistent user data structure.
