+++ package.json
@@ -21,14 +21,18 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "date-fns": "^3.6.0",
+    "i18next": "^25.2.0",
     "lucide-react": "^0.487.0",
     "lunar-javascript": "^1.7.3",
     "next": "^15.2.4",
+    "next-i18next": "^15.4.2",
     "next-themes": "^0.4.6",
     "react": "^18.3.1",
     "react-day-picker": "^8.10.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.56.3",
+    "react-i18next": "^15.5.2",
+    "react-otp-input": "^3.1.1",
     "tailwind-merge": "^3.2.0",
     "tw-animate-css": "^1.2.5",
     "zod": "^3.24.4"
