+++ public/locales/vi/index.json
@@ -1,0 +1,41 @@
+{
+  "LogIn": "Đăng nhập",
+  "Register": "Đăng ký",
+  "Sign Up with Google": "Đăng ký bằng Google",
+  "Your Email": "Email của bạn",
+  "Password": "Mật khẩu",
+  "Remember me": "Ghi nhớ đăng nhập",
+  "Forgot Password": "Quên mật khẩu",
+  "We've sent a verification code to": "Chúng tôi đã gửi mã xác minh đến",
+  "Create Your Account": "Tạo tài khoản của bạn",
+  "Verify Your Email": "Xác minh email của bạn",
+  "OR": "HOẶC",
+  "Passwords do not match": "Mật khẩu không khớp",
+  "Continue ": "Tiếp tục",
+  "Already have an account": "Bạn đã có tài khoản",
+  "Enter the 6-digit code we sent to your email": "Nhập mã gồm 6 chữ số mà chúng tôi đã gửi đến email của bạn",
+  "Didn't receive the code?": "Không nhận được mã?",
+  "Resend Code": "Gửi lại mã",
+  "Confirm Password": "Xác nhận mật khẩu",
+  "Continue": "Tiếp tục",
+  "Current Password": "Mật khẩu hiện tại",
+  "New Password": "Mật khẩu mới",
+  "Name": "Tên",
+  "Phone": "Số điện thoại",
+  "Gender": "Giới tính",
+  "Address": "Địa chỉ",
+  "Female": "Nữ",
+  "Male": "Nam",
+  "Other": "Khác",
+  "Date of Birth": "Ngày sinh nhật",
+  "Edit Profile": "Sửa thông tin",
+  "Save": "Lưu",
+  "Cancel": "Huỷ",
+  "Change avatar": "Đổi ảnh đại điện",
+  "Personal Information": "Thông tin cá nhân",
+  "Security": "Bảo mật",
+  "Password is set and hidden for security": "Mật khẩu được đặt và ẩn để bảo mật",
+  "Change Password": "Đổi mật khẩu",
+
+  "welcome": "Chào mừng bạn đến với trang web của chúng tôi!"
+}
