+++ src/components/auth/profile.tsx
@@ -1,76 +1,9 @@
-"use client";
-
-import { useState } from "react";
-import {
-  Calendar,
-  Mail,
-  MapPin,
-  Phone,
-  User,
-  Edit2,
-  Save,
-  X,
-} from "lucide-react";
-import { Button } from "@/components/ui/button";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
-import { Textarea } from "@/components/ui/textarea";
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
+import React from "react";
+import { Calendar, Mail, MapPin, Phone, User } from "lucide-react";
+import { UserType } from "../instance";
 import { format } from "date-fns";
 
-// Define the user type based on the database schema
-type UserType = {
-  name: string;
-  username: string;
-  email: string;
-  phone: string | null;
-  avatar_url: string | null;
-  gender: 0 | 1 | 2; // 0=Not specified, 1=Male, 2=Female
-  date_of_birth: string | null;
-  address: string | null;
-};
-
-export default function ProfilePage() {
-  // Sample user data
-  const [user, setUser] = useState<UserType>({
-    name: "John Doe",
-    username: "johndoe",
-    email: "john.doe@example.com",
-    phone: "+1 (555) 123-4567",
-    avatar_url: "/placeholder.svg?height=128&width=128",
-    gender: 1,
-    date_of_birth: "1990-05-15",
-    address: "123 Main Street, Apt 4B, New York, NY 10001, United States",
-  });
-
-  const [isEditing, setIsEditing] = useState(false);
-  const [editedUser, setEditedUser] = useState<UserType>(user);
-
-  const handleEdit = () => {
-    setEditedUser(user);
-    setIsEditing(true);
-  };
-
-  const handleCancel = () => {
-    setIsEditing(false);
-  };
-
-  const handleSave = () => {
-    setUser(editedUser);
-    setIsEditing(false);
-    // In a real application, you would send this data to your backend
-    console.log("Saving user data:", editedUser);
-  };
-
-  const handleChange = (field: keyof UserType, value: any) => {
-    setEditedUser((prev) => ({
-      ...prev,
-      [field]: value,
-    }));
-  };
-
+export default function Profile({ user }: { user: UserType }) {
   const getGenderLabel = (gender: 0 | 1 | 2) => {
     switch (gender) {
       case 0:
@@ -94,327 +27,49 @@ export default function ProfilePage() {
   };
 
   return (
-    <div className=" h-screen">
-      <div className="container mx-auto py-8 px-4 max-w-4xl ">
-        <div className="flex justify-between items-center mb-6">
-          <h1 className="text-3xl font-bold ">Profile</h1>
-          {!isEditing ? (
-            <Button
-              onClick={handleEdit}
-              variant="outline"
-              className="flex items-center gap-2"
-            >
-              <Edit2 className="h-4 w-4" />
-              Edit Profile
-            </Button>
-          ) : (
-            <div className="flex gap-2">
-              <Button
-                onClick={handleSave}
-                variant="default"
-                className="flex items-center gap-2"
-              >
-                <Save className="h-4 w-4" />
-                Save
-              </Button>
-              <Button
-                onClick={handleCancel}
-                variant="outline"
-                className="flex items-center gap-2"
-              >
-                <X className="h-4 w-4" />
-                Cancel
-              </Button>
-            </div>
-          )}
+    <div className="space-y-4">
+      <div className="grid grid-cols-1 gap-4">
+        <div className="flex items-center gap-2">
+          <User className="h-5 w-5 text-black dark:text-white" />
+          <div>
+            <p className="text-black dark:text-white">{user.name}</p>
+          </div>
         </div>
-
-        <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
-          {/* Profile Card */}
-          <Card className="md:col-span-1">
-            <CardHeader className="flex flex-col items-center">
-              <div className="relative mb-4">
-                <Avatar className="h-32 w-32 text-black dark:text-white">
-                  <AvatarImage
-                    src={
-                      user.avatar_url || "/placeholder.svg?height=128&width=128"
-                    }
-                    alt={user.name}
-                  />
-                  <AvatarFallback>
-                    {user.name.substring(0, 2).toUpperCase()}
-                  </AvatarFallback>
-                </Avatar>
-                {isEditing && (
-                  <div className="absolute bottom-0 right-0">
-                    <Button
-                      size="sm"
-                      variant="outline"
-                      className="rounded-full h-8 w-8 p-0 "
-                    >
-                      <Edit2 className="h-4 w-4 text-black dark:text-white" />
-                      <span className="sr-only">Change avatar</span>
-                    </Button>
-                  </div>
-                )}
-              </div>
-              <CardTitle className="text-center text-black dark:text-white">
-                {user.name}
-              </CardTitle>
-              <p className=" text-center text-black dark:text-white">
-                @{user.username}
-              </p>
-            </CardHeader>
-          </Card>
-
-          {/* Personal Information */}
-          <Card className="md:col-span-2">
-            <CardHeader>
-              <CardTitle>Personal Information</CardTitle>
-            </CardHeader>
-            <CardContent>
-              {isEditing ? (
-                <div className="space-y-4">
-                  <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                    <div className="space-y-2">
-                      <Label
-                        htmlFor="name"
-                        className="text-black dark:text-white"
-                      >
-                        Full Name
-                      </Label>
-                      <Input
-                        className="text-gray-400 dark:text-white"
-                        id="name"
-                        value={editedUser.name}
-                        onChange={(e) => handleChange("name", e.target.value)}
-                      />
-                    </div>
-                    <div className="space-y-2">
-                      <Label
-                        htmlFor="username"
-                        className="text-black dark:text-white"
-                      >
-                        Username
-                      </Label>
-                      <Input
-                        className="text-gray-400 dark:text-white"
-                        id="username"
-                        value={editedUser.username}
-                        onChange={(e) =>
-                          handleChange("username", e.target.value)
-                        }
-                      />
-                    </div>
-                  </div>
-
-                  <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                    <div className="space-y-2">
-                      <Label
-                        htmlFor="email"
-                        className="text-black dark:text-white"
-                      >
-                        Email
-                      </Label>
-                      <Input
-                        className="text-gray-400 dark:text-white"
-                        id="email"
-                        type="email"
-                        value={editedUser.email}
-                        onChange={(e) => handleChange("email", e.target.value)}
-                      />
-                    </div>
-                    <div className="space-y-2">
-                      <Label
-                        htmlFor="phone"
-                        className="text-black dark:text-white"
-                      >
-                        Phone
-                      </Label>
-                      <Input
-                        className="text-gray-400 dark:text-white"
-                        id="phone"
-                        value={editedUser.phone || ""}
-                        onChange={(e) => handleChange("phone", e.target.value)}
-                      />
-                    </div>
-                  </div>
-
-                  <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                    <div className="space-y-2">
-                      <Label
-                        htmlFor="gender"
-                        className="text-black dark:text-white"
-                      >
-                        Gender
-                      </Label>
-                      <RadioGroup
-                        value={editedUser.gender.toString()}
-                        onValueChange={(value) =>
-                          handleChange(
-                            "gender",
-                            Number.parseInt(value) as 0 | 1 | 2
-                          )
-                        }
-                        className="flex gap-4 text-black dark:text-white"
-                      >
-                        <div className="flex items-center space-x-2">
-                          <RadioGroupItem value="0" id="gender-not-specified" />
-                          <Label htmlFor="gender-not-specified">Other</Label>
-                        </div>
-                        <div className="flex items-center space-x-2">
-                          <RadioGroupItem value="1" id="gender-male" />
-                          <Label htmlFor="gender-male">Male</Label>
-                        </div>
-                        <div className="flex items-center space-x-2">
-                          <RadioGroupItem value="2" id="gender-female" />
-                          <Label htmlFor="gender-female">Female</Label>
-                        </div>
-                      </RadioGroup>
-                    </div>
-                    <div className="space-y-2 text-black dark:text-white">
-                      <Label htmlFor="date_of_birth">Date of Birth</Label>
-                      <Input
-                        className="text-gray-400 dark:text-white"
-                        id="date_of_birth"
-                        type="date"
-                        value={editedUser.date_of_birth || ""}
-                        onChange={(e) =>
-                          handleChange("date_of_birth", e.target.value)
-                        }
-                      />
-                    </div>
-                  </div>
-
-                  <div className="space-y-2">
-                    <Label
-                      htmlFor="address"
-                      className="text-black dark:text-white"
-                    >
-                      Address
-                    </Label>
-                    <Textarea
-                      className="text-black dark:text-white"
-                      id="address"
-                      value={editedUser.address || ""}
-                      onChange={(e) => handleChange("address", e.target.value)}
-                      rows={3}
-                    />
-                  </div>
-                </div>
-              ) : (
-                <div className="space-y-4">
-                  <div className="grid grid-cols-1 gap-4">
-                    <div className="flex items-center gap-2">
-                      <User className="h-5 w-5 text-black dark:text-white" />
-                      <div>
-                        <p className="text-black dark:text-white">
-                          {user.name}
-                        </p>
-                      </div>
-                    </div>
-                    <div className="flex items-center gap-2">
-                      <Mail className="h-5 w-5 text-black dark:text-white" />
-                      <div>
-                        <p className="text-black dark:text-white">
-                          {user.email}
-                        </p>
-                      </div>
-                    </div>
-                    <div className="flex items-center gap-2">
-                      <Phone className="h-5 w-5 text-black dark:text-white" />
-                      <div>
-                        <p className="text-black dark:text-white">
-                          {user.phone || "Not provided"}
-                        </p>
-                      </div>
-                    </div>
-                    <div className="flex items-center gap-2 ">
-                      <User className="h-5 w-5 text-black dark:text-white" />
-                      <div className="text-black dark:text-white">
-                        <p>{getGenderLabel(user.gender)}</p>
-                      </div>
-                    </div>
-                    <div className="flex items-center gap-2">
-                      <Calendar className="h-5 w-5 text-black dark:text-white" />
-                      <div>
-                        <p className="text-black dark:text-white">
-                          {formatDate(user.date_of_birth)}
-                        </p>
-                      </div>
-                    </div>
-                    <div className="flex items-start gap-2">
-                      <MapPin className="h-5 w-5  mt-0.5 text-black dark:text-white" />
-                      <div>
-                        <p className="whitespace-pre-wrap text-black dark:text-white">
-                          {user.address || "Not provided"}
-                        </p>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              )}
-            </CardContent>
-          </Card>
-
-          {/* Security Section */}
-          <Card className="md:col-span-3">
-            <CardHeader>
-              <CardTitle className="text-black dark:text-white">
-                Security
-              </CardTitle>
-            </CardHeader>
-            <CardContent>
-              {isEditing ? (
-                <div className="space-y-4">
-                  <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                    <div className="space-y-2">
-                      <Label
-                        htmlFor="current-password "
-                        className="text-black dark:text-white"
-                      >
-                        Current Password
-                      </Label>
-                      <Input
-                        className="text-gray-400 dark:text-white"
-                        id="current-password"
-                        type="password"
-                        placeholder="••••••••"
-                      />
-                    </div>
-                    <div className="space-y-2">
-                      <Label
-                        htmlFor="new-password "
-                        className="text-black dark:text-white"
-                      >
-                        New Password
-                      </Label>
-                      <Input
-                        className="text-gray-400 dark:text-white"
-                        id="new-password"
-                        type="password"
-                        placeholder="••••••••"
-                      />
-                    </div>
-                  </div>
-                </div>
-              ) : (
-                <div className="flex items-center gap-2">
-                  <p className="text-black dark:text-white">
-                    Password is set and hidden for security
-                  </p>
-                  <Button
-                    variant="outline"
-                    size="sm"
-                    className="ml-auto text-black dark:text-white"
-                    onClick={handleEdit}
-                  >
-                    Change Password
-                  </Button>
-                </div>
-              )}
-            </CardContent>
-          </Card>
+        <div className="flex items-center gap-2">
+          <Mail className="h-5 w-5 text-black dark:text-white" />
+          <div>
+            <p className="text-black dark:text-white">{user.email}</p>
+          </div>
+        </div>
+        <div className="flex items-center gap-2">
+          <Phone className="h-5 w-5 text-black dark:text-white" />
+          <div>
+            <p className="text-black dark:text-white">
+              {user.phone || "Not provided"}
+            </p>
+          </div>
+        </div>
+        <div className="flex items-center gap-2 ">
+          <User className="h-5 w-5 text-black dark:text-white" />
+          <div className="text-black dark:text-white">
+            <p>{getGenderLabel(user.gender)}</p>
+          </div>
+        </div>
+        <div className="flex items-center gap-2">
+          <Calendar className="h-5 w-5 text-black dark:text-white" />
+          <div>
+            <p className="text-black dark:text-white">
+              {formatDate(user.date_of_birth)}
+            </p>
+          </div>
+        </div>
+        <div className="flex items-start gap-2">
+          <MapPin className="h-5 w-5  mt-0.5 text-black dark:text-white" />
+          <div>
+            <p className="whitespace-pre-wrap text-black dark:text-white">
+              {user.address || "Not provided"}
+            </p>
+          </div>
         </div>
       </div>
     </div>
