+++ src/components/instance/index.tsx
@@ -7,3 +7,14 @@ export interface HomeProps {
     duration: number;
   }[];
 }
+
+export interface UserType {
+  name: string;
+  username: string;
+  email: string;
+  phone: string | null;
+  avatar_url: string | null;
+  gender: 0 | 1 | 2;
+  date_of_birth: string | null;
+  address: string | null;
+}
